requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
ccxt==4.2.58
pandas==2.2.1
pandas-ta==0.3.14b0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-genai==0.3.0
openai==1.14.2
httpx==0.27.0
numpy==1.26.4
aiofiles==23.2.1
